+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Increase code coverage for src/RequestForm.js by Jest/RTL tests. Refs UIREQ-1043.
 * Add functionality to handle page title on the requests list page. Refs UIREQ-1058.
 * Error in Export Hold Shelf Clearance Report. Refs UIREQ-1057.
+* Use settings/entries endpoint to get settings information. Refs UIREQ-1062.
 
 ## [9.0.1](https://github.com/folio-org/ui-requests/tree/v9.0.1) (2023-12-04)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v9.0.0...v9.0.1)